Move create_rescue_image to payload base utils
This method is used by task in the payload.base.installation so this
function should be also in the payload base python module.
